feat: do not immediately set timer to 24:59
This was a good idea, but the implementation turns out to be buggier than expected, as the underlying Timer instance is still ticking away when the viewmodel timer is paused. There may be some way to do this properly, but I haven't found it; trying to connect to Timer manually yielded no real results. Moving to v1.1.